+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `useSendMessage` now always returns the same function instance given
+  the supervisor is the same (through the `useCallback` hook).
 
 ## [0.8.2] - 2020-01-29
 ### Added